rename rust module evm_runner
