+++ runtime/common/src/lib.rs
@@ -19,11 +19,11 @@
 use sp_core::H160;
 
 mod apis;
-mod evm_runner;
 mod impl_moonbeam_xcm_call;
 mod impl_moonbeam_xcm_call_tracing;
 mod impl_on_charge_evm_transaction;
 mod impl_self_contained_call;
+mod impl_xcm_evm_runner;
 pub mod migrations;
 
 frame_support::parameter_types! {
